+++ app/Console/HaveAccess.php
@@ -36,8 +36,8 @@ trait HaveAccess
      */
     private function haveAccess(): bool
     {
-        $uri   = (string)config('csv_importer.uri');
-        $token     = (string)config('csv_importer.access_token');
+        $uri     = (string) config('csv_importer.uri');
+        $token   = (string) config('csv_importer.access_token');
         $request = new SystemInformationRequest($uri, $token);
 
         $request->setVerify(config('csv_importer.connection.verify'));