+++ app/Http/Controllers/TokenController.php
@@ -25,12 +25,16 @@ namespace App\Http\Controllers;
 
 use GrumpyDictator\FFIIIApiSupport\Exceptions\ApiHttpException;
 use GrumpyDictator\FFIIIApiSupport\Request\SystemInformationRequest;
-use GrumpyDictator\FFIIIApiSupport\Response\SystemInformationResponse;
+use GuzzleHttp\Client;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Http\Request;
 use Illuminate\Routing\Redirector;
 use Illuminate\View\View;
+use InvalidArgumentException;
+use Log;
+use Str;
 
 /**
  * Class TokenController
@@ -38,15 +42,101 @@ use Illuminate\View\View;
 class TokenController extends Controller
 {
     /**
+     * @param Request $request
+     * @return \Illuminate\Contracts\Foundation\Application|Factory|RedirectResponse|Redirector|View
+     */
+    public function index(Request $request)
+    {
+        $pageTitle = 'CSV importer';
+        Log::debug(sprintf('Now at %s', __METHOD__));
+        $configToken = (string) config('csv_importer.access_token');
+        $clientId    = (int) config('csv_importer.client_id');
+        $baseURL     = (string) config('csv_importer.uri');
+
+        Log::info('The following configuration information was found:');
+        Log::info(sprintf('Personal Access Token: "%s". (limited to 25 chars if present)', substr($configToken, 0, 25)));
+        Log::info(sprintf('Client ID            : "%s".', $clientId));
+        Log::info(sprintf('Base URL             : "%s".', $baseURL));
+
+        // Option 1: access token and url are present:
+        if ('' !== $configToken && '' !== $baseURL) {
+            Log::debug(sprintf('Found personal access token + URL "%s" in config, set cookie and return to index.', $baseURL));
+
+            // set cookies.
+            $cookies = [
+                cookie('access_token', $configToken),
+                cookie('base_url', $baseURL),
+                cookie('refresh_token', ''),
+            ];
+
+            return redirect(route('index'))->withCookies($cookies);
+        }
+
+        // Option 2: client ID + base URL.
+        if (0 !== $clientId && '' !== $baseURL) {
+            Log::debug(sprintf('Found client ID "%d" + URL "%s" in config, redirect to Firefly III for permission.', $clientId, $baseURL));
+
+            // send user to vanity URL!
+            if ('' !== (string) config('csv_importer.vanity_uri')) {
+                $baseURL = config('csv_importer.vanity_uri');
+            }
+
+            return $this->redirectForPermission($request, $baseURL, $clientId);
+        }
+
+        // Option 3: either is empty, ask for client ID and/or base URL:
+        $baseURL  = config('csv_importer.uri');
+        $clientId = 0 === $clientId ? '' : $clientId;
+        return view('token.client_id', compact('baseURL', 'clientId', 'pageTitle'));
+    }
+
+
+    /**
+     * @param Request $request
+     */
+    public function submitClientId(Request $request)
+    {
+        Log::debug(sprintf('Now at %s', __METHOD__));
+        $data = $request->validate(
+            [
+                'client_id' => 'required|numeric|min:1|max:65536',
+                'base_url'  => 'url',
+            ]);
+        Log::debug('Submitted data: ', $data);
+
+        if (true === config('csv_importer.expect_secure_uri') && 'https://' !== substr($data['base_url'], 0, 8)) {
+            $request->session()->flash('secure_url','URL must start with https://');
+            return redirect(route('token.index'));
+        }
+
+
+        $data['client_id'] = (int) $data['client_id'];
+
+        // grab base URL from config first, otherwise from submitted data:
+        $baseURL = config('csv_importer.uri');
+        if ('' !== (string) config('csv_importer.vanity_uri')) {
+            $baseURL = config('csv_importer.vanity_uri');
+        }
+        if (array_key_exists('base_url', $data) && '' !== $data['base_url']) {
+            $baseURL = $data['base_url'];
+        }
+        $baseURL = rtrim($baseURL, '/');
+
+        // return request for permission:
+        return $this->redirectForPermission($request, $baseURL, $data['client_id']);
+    }
+
+    /**
      * Check if the Firefly III API responds properly.
      *
      * @return JsonResponse
      */
-    public function doValidate(): JsonResponse
+    public function doValidate(Request $request): JsonResponse
     {
+        Log::debug(sprintf('Now at %s', __METHOD__));
         $response = ['result' => 'OK', 'message' => null];
-        $uri   = (string)config('csv_importer.uri');
-        $token   = (string)config('csv_importer.access_token');
+        $uri      = (string) $request->cookie('base_url');
+        $token    = (string) $request->cookie('access_token');
         $request  = new SystemInformationRequest($uri, $token);
 
         $request->setVerify(config('csv_importer.connection.verify'));
@@ -55,7 +145,7 @@ class TokenController extends Controller
         try {
             $result = $request->get();
         } catch (ApiHttpException $e) {
-            return ['result' => 'NOK', 'message' => $e->getMessage()];
+            return response()->json(['result' => 'NOK', 'message' => $e->getMessage()]);
         }
         // -1 = OK (minimum is smaller)
         // 0 = OK (same version)
@@ -75,49 +165,83 @@ class TokenController extends Controller
     }
 
     /**
-     * Same thing but not over JSON.
-     *
-     * @return Factory|RedirectResponse|Redirector|View
+     * @param Request $request
      */
-    public function index()
+    public function callback(Request $request)
     {
-        $uri          = (string)config('csv_importer.uri');
-        $token        = (string)config('csv_importer.access_token');
-        $request      = new SystemInformationRequest($uri, $token);
-        $errorMessage = 'No error message.';
-        $isError      = false;
-        $result       = null;
-        $compare      = 1;
-        $minimum      = '';
+        Log::debug(sprintf('Now at %s', __METHOD__));
+        $state        = (string) $request->session()->pull('state');
+        $codeVerifier = (string) $request->session()->pull('code_verifier');
+        $clientId     = (int) $request->session()->pull('form_client_id');
+        $baseURL      = (string) $request->session()->pull('form_base_url');
+        $code         = $request->get('code');
 
-        $request->setVerify(config('csv_importer.connection.verify'));
-        $request->setTimeOut(config('csv_importer.connection.timeout'));
+        throw_unless(
+            strlen($state) > 0 && $state === $request->state,
+            InvalidArgumentException::class
+        );
+        $finalURL = sprintf('%s/oauth/token', $baseURL);
+        $params   = [
+            'form_params' => [
+                'grant_type'    => 'authorization_code',
+                'client_id'     => $clientId,
+                'redirect_uri'  => route('token.callback'),
+                'code_verifier' => $codeVerifier,
+                'code'          => $code,
+            ],
+        ];
+        Log::debug('State is valid!');
+        Log::debug('Params for access token', $params);
+        Log::debug(sprintf('Will contact "%s" for a token.', $finalURL));
 
-        try {
-            /** @var SystemInformationResponse $result */
-            $result = $request->get();
-        } catch (ApiHttpException $e) {
-            $errorMessage = $e->getMessage();
-            $isError      = true;
-        }
-        // -1 = OK (minimum is smaller)
-        // 0 = OK (same version)
-        // 1 = NOK (too low a version)
-        if (false === $isError) {
-            $minimum = config('csv_importer.minimum_version');
-            $compare = version_compare($minimum, $result->version);
-        }
-        if (false === $isError && 1 === $compare) {
-            $errorMessage = sprintf('Your Firefly III version %s is below the minimum required version %s', $result->version, $minimum);
-            $isError      = true;
-        }
 
-        if (false === $isError) {
-            return redirect(route('index'));
-        }
-        $pageTitle = 'Token error';
+        $response = (new Client)->post($finalURL, $params);
+        $data     = json_decode((string) $response->getBody(), true);
 
-        return view('token.index', compact('errorMessage', 'pageTitle'));
+        Log::debug('Response', $data);
+
+        // set cookies.
+        $cookies = [
+            cookie('access_token', (string) $data['access_token']),
+            cookie('base_url', $baseURL),
+            cookie('refresh_token', (string) $data['refresh_token']),
+        ];
+        Log::debug(sprintf('Return redirect with cookies to "%s"', route('index')));
+
+        return redirect(route('index'))->withCookies($cookies);
+    }
+
+    /**
+     * @param Request $request
+     * @param string  $baseURL
+     * @param int     $clientId
+     * @return RedirectResponse
+     */
+    private function redirectForPermission(Request $request, string $baseURL, int $clientId): RedirectResponse
+    {
+        $state        = Str::random(40);
+        $codeVerifier = Str::random(128);
+        $request->session()->put('state', $state);
+        $request->session()->put('code_verifier', $codeVerifier);
+        $request->session()->put('form_client_id', $clientId);
+        $request->session()->put('form_base_url', $baseURL);
+
+        $codeChallenge = strtr(rtrim(base64_encode(hash('sha256', $codeVerifier, true)), '='), '+/', '-_');
+        $params        = [
+            'client_id'             => $clientId,
+            'redirect_uri'          => route('token.callback'),
+            'response_type'         => 'code',
+            'scope'                 => '',
+            'state'                 => $state,
+            'code_challenge'        => $codeChallenge,
+            'code_challenge_method' => 'S256',
+        ];
+        $query         = http_build_query($params);
+        $finalURL      = sprintf('%s/oauth/authorize?', $baseURL);
+        Log::debug('Query parameters are', $params);
+        Log::debug(sprintf('Now redirecting to "%s" (params omitted)', $finalURL));
+
+        return redirect($finalURL . $query);
     }
 
 }