+++ app/Services/CSV/Mapper/TransactionCurrencies.php
@@ -24,6 +24,7 @@ declare(strict_types=1);
 namespace App\Services\CSV\Mapper;
 
 use App\Exceptions\ImportException;
+use App\Support\Token;
 use GrumpyDictator\FFIIIApiSupport\Exceptions\ApiHttpException;
 use GrumpyDictator\FFIIIApiSupport\Model\TransactionCurrency;
 use GrumpyDictator\FFIIIApiSupport\Request\GetCurrenciesRequest;
@@ -44,8 +45,8 @@ class TransactionCurrencies implements MapperInterface
     public function getMap(): array
     {
         $result = [];
-        $uri    = (string) config('csv_importer.uri');
-        $token  = (string) config('csv_importer.access_token');
+        $uri    = Token::getURL();
+        $token  = Token::getAccessToken();
 
         $request = new GetCurrenciesRequest($uri, $token);
         $request->setVerify(config('csv_importer.connection.verify'));